bma-backend/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.2.1
autopep8==2.0.4
backcall==0.2.0
beautifulsoup4==4.12.2
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coards==1.0.5
colorama==0.4.6
comm==0.1.4
contourpy==1.0.7
cssselect2==0.7.0
cx-Oracle==8.3.0
cycler==0.11.0
debugpy==1.6.7.post1
decorator==5.1.1
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
docopt==0.6.2
et-xmlfile==1.1.0
executing==1.2.0
Extract==1.0
fonttools==4.40.0
gobject==0.1.0
html5lib==1.1
idna==3.4
ipykernel==6.25.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.23.4
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==23.1
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.3
pure-eval==0.2.2
pycodestyle==2.11.0
pycparser==2.21
pydap==3.4.1
pydyf==0.6.0
Pygments==2.16.1
PyJWT==2.4.0
pyparsing==3.0.9
pyphen==0.14.0
pythainlp==4.0.2
python-dateutil==2.8.2
pytz==2022.5
pywin32==306
pyzmq==25.1.1
requests==2.28.1
seaborn==0.12.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
tzdata==2022.5
urllib3==1.26.12
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
WebOb==1.8.7
wfastcgi==3.0.0
xmltodict==0.13.0
zopfli==0.2.2
